feat: Implement user service management and order history functionality

+++ controllers/admin/auth.controller.js
@@ -12,7 +12,7 @@ import { generateAccessToken } from "../../utils/generateToken.js";
 export const Login = async (req, res) => {
     try {
         const { username, password } = req.body;
-       
+
         const admin = await AdminModel.findOne({ username });
         if (!admin) {
             return res.status(400).json(new APIError("Admin not found", 400))
@@ -22,10 +22,10 @@ export const Login = async (req, res) => {
         if (!match) {
             return res.status(400).json(new APIError("Invalid credentials", 400))
         }
-        
-        const Token = generateAccessToken(admin._id,admin.role);
 
-        return res.status(200).json(new APIResponse("Admin Login Successfully!", 200, {Token:Token}))
+        const Token = generateAccessToken(admin._id, admin.role);
+
+        return res.status(200).json(new APIResponse("Admin Login Successfully!", 200, { Token: Token }))
     } catch (error) {
         console.log(error);
 
@@ -46,14 +46,14 @@ export const UpdatePassowrd = async (req, res) => {
         if (!admin) {
             return res.status(404).json(new APIError("Admin Not Found", 404))
         }
-       
-        
+
+
         const match = await bcrypt.compare(oldPassword, admin.password);
         if (!match) {
             return res.status(400).json(new APIError("Invalid Credential", 400))
         }
 
-        admin.password =await bcrypt.hash(newPassword,10);
+        admin.password = await bcrypt.hash(newPassword, 10);
         await admin.save();
         return res.status(200).json(new APIResponse("Admin Password Updated Successfully!", 200))
     } catch (error) {
@@ -76,7 +76,7 @@ export const UpdateAdminProfile = async (req, res) => {
         if (!admin) {
             return res.status(404).json(new APIError("Admin Not Found", 404));
         }
-       
+
         admin.username = username;
         await admin.save();
         return res.status(200).json(new APIResponse("Admin Password Updated Successfully!", 200))
@@ -106,66 +106,101 @@ export const GetAdminProfile = async (req, res) => {
 
 //USER ROUTES =========================================================
 
-
-export const GetUsersList=async (req, res) => {
+export const CreateUser = async (req, res) => {
     try {
-        const usersList=await userModel.find({});
+        const { userName, mobileNumber, password } = req.body;
 
-        if(usersList.length==0){
-            return res.status(400).json(new APIError("Users List Empty",400));
+        if (!userName || !mobileNumber || !password) {
+            return res.status(400).json(new APIError("Please Fill All Required Fileds", 400))
         }
-  
-        return res.status(200).json(new APIResponse("Users List :", 200,usersList))
+        
+        const isExist = await userModel.findOne({ userName: userName, mobileNumber: mobileNumber });
+        
+        if (isExist) {
+            return res.status(400).json(new APIError("User with this username or mobile number already exist", 400))
+        }
+         
+
+        await userModel.create({
+            userName: userName,
+            mobileNumber: mobileNumber,
+            password: password,
+            role: "user",
+            refreshToken: "",
+            content: "",
+            status: req.body.status ?? false,
+            block: req.body.block ?? false,
+            deleted: req.body.deleted ?? false,
+            isSetupDone: req.body.isSetupDone ?? false,
+            adminId:req.user.id
+        })
+
+        return res.status(200).json(new APIResponse("User Created Successfully!", 200))
     } catch (error) {
         console.log(error);
         return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
 
-export const GetUserById=async (req, res) => {
+export const GetUsersList = async (req, res) => {
     try {
-         const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
+        const usersList = await userModel.find({adminId:req.user.id});
 
-         const user=await userModel.findOne({_id:userId});
+        if (usersList.length == 0) {
+            return res.status(400).json(new APIError("Users List Empty", 400));
+        }
 
-         if(!user){
-            return res.status(404).json(new APIError("User Not Found",404));
-         }
-
-        return res.status(200).json(new APIResponse("User Profile :", 200,user))
+        return res.status(200).json(new APIResponse("Users List :", 200, usersList))
     } catch (error) {
         console.log(error);
         return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
 
-export const UpdateUserPassword=async (req, res) => {
+export const GetUserById = async (req, res) => {
     try {
-         const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
-         const { oldPassword, newPassword } = req.body;
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
+
+        const user = await userModel.findOne({ _id: userId,adminId:req.user.id });
+
+        if (!user) {
+            return res.status(404).json(new APIError("User Not Found", 404));
+        }
+
+        return res.status(200).json(new APIResponse("User Profile :", 200, user))
+    } catch (error) {
+        console.log(error);
+        return res.status(500).json(new APIError("Error :" + error, 500))
+    }
+}
+
+export const UpdateUserPassword = async (req, res) => {
+    try {
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
+        const { oldPassword, newPassword } = req.body;
         if (!oldPassword || !newPassword) {
             return res.status(400).json(new APIError("Old Password Or New Password Not Found", 400));
         }
-         const user=await userModel.findOne({_id:userId});
+        const user = await userModel.findOne({ _id: userId,adminId:req.user.id });
 
-         if(!user){
-            return res.status(404).json(new APIError("User Not Found",404));
-         }
-         console.log();
-         
+        if (!user) {
+            return res.status(404).json(new APIError("User Not Found", 404));
+        }
+        console.log();
+
         //  const match= await bcrypt.compare(oldPassword,user.password);
-         if(oldPassword!==user.password){
-            return res.status(400).json(new APIError("Invalid Credential",400));
-         }
+        if (oldPassword !== user.password) {
+            return res.status(400).json(new APIError("Invalid Credential", 400));
+        }
         //  const hashedPassword=await bcrypt.hash(newPassword,10);
-         user.password=newPassword;
-         await user.save();
+        user.password = newPassword;
+        await user.save();
         return res.status(200).json(new APIResponse("User Password Updated Successfully!", 200));
     } catch (error) {
         console.log(error);
@@ -173,18 +208,18 @@ export const UpdateUserPassword=async (req, res) => {
     }
 }
 
-export const DeleteUser=async (req, res) => {
+export const DeleteUser = async (req, res) => {
     try {
-         const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
 
-         const isDeleted=await userModel.findOneAndDelete({_id:userId});
+        const isDeleted = await userModel.findOneAndDelete({ _id: userId ,adminId:req.user.id});
 
-         if(!isDeleted){
-            return res.status(400).json(new APIError("Error Occure When Deleting User",400));
-         }
+        if (!isDeleted) {
+            return res.status(400).json(new APIError("Error Occure When Deleting User", 400));
+        }
 
         return res.status(200).json(new APIResponse("User Deleted Successfully!", 200))
     } catch (error) {
@@ -193,82 +228,83 @@ export const DeleteUser=async (req, res) => {
     }
 }
 
-export const UpdateUserStatus=async(req,res)=>{
+export const UpdateUserStatus = async (req, res) => {
     try {
-        const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
 
-         const user=await userModel.findOne({_id:userId});
+        const user = await userModel.findOne({ _id: userId,adminId:req.user.id });
         console.log(user);
-        
-         if(!user){
-            return res.status(400).json(new APIError("User Not Found",400));
-         }
-         user.status=req.body.status??user.status;
 
-         await user.save();
+        if (!user) {
+            return res.status(400).json(new APIError("User Not Found", 400));
+        }
+        user.status = req.body.status ?? user.status;
+
+        await user.save();
         return res.status(200).json(new APIResponse("User Status Updated Successfully!", 200))
     } catch (error) {
-         return res.status(500).json(new APIError("Error :" + error, 500))
+        return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
 
-export const BlockUser=async(req,res)=>{
+export const BlockUser = async (req, res) => {
     try {
-        const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
 
-         const user=await userModel.findOne({_id:userId});
+        const user = await userModel.findOne({ _id: userId,adminId:req.user.id });
 
-         if(!user){
-            return res.status(400).json(new APIError("User Not Found",400));
-         }
-         
-          user.block=req.body.block??user.block;
-         await user.save();
+        if (!user) {
+            return res.status(400).json(new APIError("User Not Found", 400));
+        }
+
+        user.block = req.body.block ?? user.block;
+        await user.save();
         return res.status(200).json(new APIResponse("User Blocked Successfully!", 200))
     } catch (error) {
-         return res.status(500).json(new APIError("Error :" + error, 500))
+        return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
 
-export const DeletedUser=async(req,res)=>{
+export const DeletedUser = async (req, res) => {
     try {
-        const userId=req.params.id;
-         if(!userId){
-            return res.status(400).json(new APIError("User Id Missing",400));
-         }
+        const userId = req.params.id;
+        if (!userId) {
+            return res.status(400).json(new APIError("User Id Missing", 400));
+        }
 
-         const user=await userModel.findOne({_id:userId});
+        const user = await userModel.findOne({ _id: userId ,adminId:req.user.id});
 
-         if(!user){
-            return res.status(400).json(new APIError("User Not Found",400));
-         }
-         
-         user.deleted=req.body.deleted??user.deleted;
-         await user.save();
+        if (!user) {
+            return res.status(400).json(new APIError("User Not Found", 400));
+        }
+
+        user.deleted = req.body.deleted ?? user.deleted;
+        await user.save();
         return res.status(200).json(new APIResponse("User Deleted Successfully!", 200))
     } catch (error) {
-         return res.status(500).json(new APIError("Error :" + error, 500))
+        return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
 
-export const GetAllUserNameList=async (req, res) => {
+export const GetAllUserNameList = async (req, res) => {
     try {
-        const usersNameList=await userModel.find({},{userName:1});
+        const usersNameList = await userModel.find({adminId:req.user.id}, { userName: 1 });
 
-        if(usersNameList.length==0){
-            return res.status(400).json(new APIError("Users List Empty",400));
+        if (usersNameList.length == 0) {
+            return res.status(400).json(new APIError("Users List Empty", 400));
         }
-  
-        return res.status(200).json(new APIResponse("Users Names List :", 200,usersNameList))
+
+        return res.status(200).json(new APIResponse("Users Names List :", 200, usersNameList))
     } catch (error) {
         console.log(error);
         return res.status(500).json(new APIError("Error :" + error, 500))
     }
 }
+
